projects/H-031/requirements.txt
numpy==2.3.5
torch==2.9.1
streamlit==1.52.2
matplotlib==3.10.8
Pillow==12.0.0
scikit-learn==1.8.0
